% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_in}
\alias{arrange_in}
\title{Arrange data frames within chain by the column values}
\usage{
arrange_in(.chain, .dataset, ...)
}
\arguments{
\item{.chain}{A chain.}

\item{.dataset}{The name of the data frame or selection verbs.}

\item{...}{Similar to \code{dplyr::arrange}.}
}
\description{
Arrange data frames within chain by the column values
}
