% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainr-package.R
\docType{package}
\name{chainr-package}
\alias{chainr}
\alias{chainr-package}
\title{chainr: Link data frames as a chain}
\description{
This is a tiny package provides tools to make a chain (a list of related data
 frames) and manipulate using familiar dplyr verbs made for a chain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/chainr}
  \item Report bugs at \url{https://github.com/emitanaka/chainr/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
